feat: create home page component, move header and main components from app to home page, update header styles, remove error button
