+++ src/app/App.module.css
@@ -1,6 +1,4 @@
 .app {
-  display: flex;
-  flex-direction: column;
-  gap: 40px;
-  padding-bottom: 40px;
+  display: grid;
+  grid-template-columns: auto 1fr;
 }
