+++ src/app/App.tsx
@@ -1,35 +1,10 @@
 import styles from './App.module.css';
-import { Header, Main } from '../components';
-import { People } from '../serviсes/api.props';
-import { fetchApi } from '../serviсes/api';
-import { useEffect, useState } from 'react';
+import { HomePage } from '../pages';
 
 function App() {
-  const [searchData, setSearchData] = useState<People[]>([]);
-  const [loading, setLoading] = useState<boolean>(false);
-
-  useEffect(() => {
-    const savedSearchString = localStorage.getItem('searchString');
-
-    if (savedSearchString) {
-      handleSearch(savedSearchString);
-    } else {
-      handleSearch('');
-    }
-  }, []);
-
-  const handleSearch = (searchString: string) => {
-    setLoading(true);
-    fetchApi(searchString).then((data) => {
-      setSearchData(data.results);
-      setLoading(false);
-    });
-  };
-
   return (
     <div className={styles['app']}>
-      <Header handleSearch={handleSearch} />
-      <Main searchData={searchData} loading={loading} />
+      <HomePage />
     </div>
   );
 }
