+++ src/components/header/Header.module.css
@@ -1,11 +1,11 @@
 .header {
   display: flex;
   justify-content: space-between;
+  flex-direction: column;
   background-color: #141414;
   border-radius: 2px;
   padding: 20px;
   gap: 10px;
-  flex-wrap: wrap;
 }
 
 .description {
@@ -13,11 +13,12 @@
   font-weight: 400;
   line-height: 1.5;
   color: rgba(255, 255, 255, 0.87);
-  font-family: 'Roboto', sans-serif;
+  font-family: "Roboto", sans-serif;
 }
 
 .form {
-  display: flex;
+  display: grid;
+  grid-template-columns: 1fr auto;
   gap: 10px;
 }
 
@@ -26,7 +27,7 @@
   background-color: #f8b600;
   border: none;
   border-radius: 2px;
-  font-family: 'Roboto', sans-serif;
+  font-family: "Roboto", sans-serif;
   font-weight: bold;
   color: #161616;
   cursor: pointer;
