+++ src/components/header/Header.tsx
@@ -5,7 +5,6 @@ import { Button, SearchInput } from '..';
 
 function Header({ handleSearch }: HeaderProps) {
   const [searchString, setSearchString] = useState<string>(localStorage.getItem('searchString') || '');
-  const [error, setError] = useState<boolean>(false);
 
   const handleInputChange = (value: string) => {
     setSearchString(value);
@@ -17,14 +16,6 @@ function Header({ handleSearch }: HeaderProps) {
     handleSearch(searchString.trim());
   };
 
-  const throwError = () => {
-    setError(true);
-  };
-
-  if (error) {
-    throw new Error('Ops! Something went wrong!');
-  }
-
   return (
     <header className={styles['header']}>
       <div className={styles['logotype']}>
@@ -34,7 +25,6 @@ function Header({ handleSearch }: HeaderProps) {
         <SearchInput value={searchString} onChange={handleInputChange} />
         <Button className={styles['search-button']} text="Search" />
       </form>
-      <Button className={styles['error-button']} onClick={throwError} text="Throw error" />
     </header>
   );
 }
