+++ src/components/search-input/SearchInput.module.css
@@ -2,8 +2,7 @@
   border: none;
   border-radius: 5px;
   padding: 0px 15px;
-  width: 250px;
-  font-family: 'Roboto', sans-serif;
+  font-family: "Roboto", sans-serif;
   font-weight: bold;
 
   &:focus {
