+++ src/pages/home/HomePage.module.css
@@ -1,0 +1,2 @@
+.aside {
+}
